# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API",
    "version": "1.0",
    "description": "Auth API with Supabase — signup, login, logout, and protected routes"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Sign up a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing or invalid input" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns access_token and refresh_token" },
          "400": { "description": "Missing input" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public info — no auth needed",
        "responses": { "200": { "description": "Public data" } }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Authenticated dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Welcome message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    }
  }
}